Modularised done Global behaviour is placed in domUtils.js and navbar.js. Testing requiredgit add . Does this mean I can move forward with rest of site? Keep inputting to the things to do file as is handy reference and keeps project on point...
