+++ js/navbar.js
@@ -1,9 +1,7 @@
 // navbar.js
-// import { showWip, hideWip } from "./messagesAndForms";
-// Handles the sticky navbar logic
-
-// Main function to initialize the navbar
-import { getDomElements, getEndTopY } from "./domUtils.js";
+import { getDomElements, getEndTopY, getThumbWidthWithoutMargin, 
+  updateDimensionsNoMargins, collectThumbs, showStatementContact,
+  showForm  } from "./domUtils.js";
 
 
 const domElements = getDomElements(); 
@@ -63,17 +61,13 @@ export function setupDynamicLinks() {
   });
 }
 
-// navbar.js - Centralized Navbar Behavior
-
-let hoverTimeout; // Used for delayed dropdown hiding
+let hoverTimeout; 
 
 export function setupNavbarEvents(domElements) {
   if (!domElements) {
     console.error("Navbar elements not found!");
     return;
   }
-
-  console.log("✅ domElements before setupDropdownHover:", domElements);
   // Setup dropdown hover behaviour
   setupDropdownHover(domElements.navSoftware, domElements.navDropMenuSoftware, domElements.thumbSoft);
   setupDropdownHover(domElements.navMotion, domElements.navDropMenuMotion, domElements.thumbMot);
@@ -92,13 +86,6 @@ export function setupNavbarEvents(domElements) {
   });
 }
 
-// export function showStatementContact() {
-//   updateDimensionsNoMargins();
-//   document.getElementById("statementContact").style.display = "block";
-//   document.getElementById("contactForm").style.display = "none";
-//   collectThumbs();
-// }
-
 // 🔹 Handles dropdown hover behavior
 function setupDropdownHover(navItem, dropdownMenu, thumbElements) {
   if (!navItem || !dropdownMenu) {
@@ -109,13 +96,13 @@ function setupDropdownHover(navItem, dropdownMenu, thumbElements) {
   navItem.addEventListener("mouseenter", function () {
     showDropMenu(dropdownMenu);
     navItem.classList.add("active");
-    if (thumbElements) thumbElements.classList.add("active"); // ✅ Only add if exists
+    if (thumbElements) thumbElements.classList.add("active"); 
   });
 
   navItem.addEventListener("mouseleave", function () {
     delayedHide(dropdownMenu);
     navItem.classList.remove("active");
-    if (thumbElements) thumbElements.classList.remove("active"); // ✅ Only remove if exists
+    if (thumbElements) thumbElements.classList.remove("active"); 
   });
 
   dropdownMenu.addEventListener("mouseenter", cancelHide);
@@ -176,23 +163,10 @@ export function autoScrollNow() {
 
 // Displays WIP message
 export function showWip() {
-  // // const domElements = getDomElements(); // ✅ Ensure we have latest elements
-
-  // console.log("✅ domElements in showWip:", domElements); // ✅ Debugging
-
-  // if (!domElements.thumbElements || domElements.thumbElements.length === 0) {
-  //   console.error("❌ Error: thumbElements is missing or empty!", domElements.thumbElements);
-  //   return;
-  // }
-  // updateWIPDimensions(getEndTopY(), domElements.thumbElements);
-  // updateDimensionsNoMargins();
-  // document.getElementById("wip").style.display = "block";
-
   const domElements = getDomElements();
-  console.log("✅ domElements in showWip:", domElements);
 
   if (!domElements.thumbElements || domElements.thumbElements.length === 0) {
-    console.error("❌ Error: thumbElements is missing or empty!", domElements.thumbElements);
+    console.error("Error: thumbElements is missing or empty!", domElements.thumbElements);
     return;
   }
 
@@ -202,16 +176,8 @@ export function showWip() {
   updateWIPDimensions(endTopY, domElements.thumbElements);
   updateDimensionsNoMargins();
   document.getElementById("wip").style.display = "block";
-
-
   collectThumbs();
 }
-//   updateWIPDimensions(getEndTopY(), thumbElements);
-//   updateDimensionsNoMargins();
-//   document.getElementById("wip").style.display = "block";
-//   document.getElementById("contactForm").style.display = "none";
-// }
-
 
 // Hides WIP message
 export function hideWip() {
@@ -244,32 +210,6 @@ export function updateWIPDimensions(endTopY, thumbElements) {
     return;
   }
 
-  // const thumbWidthWithoutMargin = getThumbWidthWithoutMargin();
-
-  // // Calculate new width and height for the modal box
-  // const newWidth = Math.max(thumbWidthWithoutMargin * 2 + 4, 300);
-  // const newHeight = newWidth; // Assuming we want a square wip
-
-  // // Update modal dimensions
-  // wip.style.width = `${newWidth}px`;
-  // wip.style.height = `${newHeight}px`;
-
-  // // Calculate the center of the screen
-  // const centerX = window.innerWidth / 2 + 6; //refining positioning
-
-  // // Calculate the new left position to center the modal box
-  // const newLeft = centerX - newWidth / 2;
-
-  // // Use the passed endTopY for the new top position
-  // const newTop = endTopY + 12.5; //  works for alignment on y axis
-
-  // // Update modal position
-  // wip.style.position = "absolute";
-  // wip.style.left = `${newLeft}px`;
-  // wip.style.top = `${newTop}px`;
-
-  // // modalBox.style.display = "block";
-
   const thumbWidthWithoutMargin = getThumbWidthWithoutMargin();
   const newWidth = Math.max(thumbWidthWithoutMargin * 2 + 4, 300);
   const newHeight = newWidth; // Assuming a square WIP
@@ -288,22 +228,6 @@ export function updateWIPDimensions(endTopY, thumbElements) {
   console.log("✅ WIP dimensions updated with thumbElements.");
 }
 
-// export function getThumbWidthWithoutMargin() {
-//   // const domElements = getDomElements();
-//   const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
-//   const thumbWidth = parseFloat(computedStyle.getPropertyValue("width"));
-
-//   return thumbWidth;
-// }
-
-export function getThumbMargin() {
-  const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
-  const thumbMargin = parseFloat(
-    computedStyle.getPropertyValue("margin-right")
-  );
-  return thumbMargin;
-}
-
 export function getThumbWidthWithMargin() {
   const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
   const thumbWidth = parseFloat(computedStyle.getPropertyValue("width"));
@@ -314,210 +238,12 @@ export function getThumbWidthWithMargin() {
   return thumbWidth + thumbMargin * 2;
 }
 
-// function getComputedStyleValue(element, property) {
 
-//   return parseInt(window.getComputedStyle(element).getPropertyValue(property));
-// }
-
-// let screenWidthHalved = svg.viewBox.baseVal.width / 2;
-// let screenHeightHalved = svg.viewBox.baseVal.height / 2;
-// export let endLeftX, endRightX, endBottomY; 
-// let thumbWidth = getComputedStyleValue(
-//   document.querySelector(".thumbShape"),
-//   "width"
-// );
 
 export const landscapeMediaQuery = window.matchMedia(
   "(orientation: landscape) and (max-width: 991.98px) and (max-height: 600px)"
 );
 
-// export function updateModalDimensions(endTopY) {
-//   const modalBox = document.querySelector(".modalbox .box");
-
-//   if (!modalBox || !domElements.thumbElements[0]) return;
-
-//   const thumbWidthWithoutMargin = getThumbWidthWithoutMargin();
-
-//   // Calculate new width and height for the modal box
-//   const newWidth = Math.max(thumbWidthWithoutMargin * 2 + 4, 300);
-//   const newHeight = newWidth; // Assuming we want a square modal
-
-//   // Update modal dimensions
-//   modalBox.style.width = `${newWidth}px`;
-//   modalBox.style.height = `${newHeight}px`;
-
-//   // Calculate the center of the screen
-//   const centerX = window.innerWidth / 2 + 6; //refining positioning
-
-//   // Calculate the new left position to center the modal box
-//   const newLeft = centerX - newWidth / 2;
-
-//   // Use the passed endTopY for the new top position
-//   const newTop = getEndTopY() + 12.5; //  works for alignment on y axis
-
-//   // Update modal position
-//   modalBox.style.position = "absolute";
-//   modalBox.style.left = `${newLeft}px`;
-//   modalBox.style.top = `${newTop}px`;
-
-//   // modalBox.style.display = "block";
-// }
 
 
-
-// Contact form sizing and positioning
-export function formControl(endTopY) {
-  const contactForm = document.querySelector(".formDiv#contactForm");
-
-  if (!contactForm) return;
-
-  const computedStyleForm = window.getComputedStyle(contactForm);
-  const formWidth = parseFloat(computedStyleForm.getPropertyValue("width"));
-  const formHeight = parseFloat(computedStyleForm.getPropertyValue("height"));
-
-  // Calculate form position
-  const formX = window.innerWidth / 2 - formWidth / 2 + 6.75;
-  const formY = getEndTopY() + 12; //+ 190;
-
-  // Update modal position
-  contactForm.style.position = "absolute";
-  contactForm.style.left = `${formX}px`;
-  contactForm.style.top = `${formY}px`;
-}
-
-
-// export function updateDimensions() {
-//    thumbWidth = Math.min(300, window.innerWidth / 6);
-//   screenWidthHalved = window.innerWidth / 2;
-//   screenHeightHalved = window.innerHeight / 2;
-
-//   endLeftX = screenWidthHalved - totalThumbWidth;
-
-//   updateEndTopY(); // ✅ Update the value first
-//   const updatedEndTopY = getEndTopY(); // ✅ Retrieve the updated value
-
-//   if (updatedEndTopY === undefined) {
-//     console.error("Error: getEndTopY() returned undefined!");
-//     return; // Stop execution if the value is not set
-//   }
-
-//   const endTopY = updatedEndTopY; // ✅ Use a local variable, do not redeclare globally
-
-//   endRightX = screenWidthHalved;
-//   updateEndBottomY();
-// }
-
-// export function spaceoutThumbs() {
-//   gsap.to("#software", {
-//     x: endLeftX,
-//     y: getEndTopY(),
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#photography", {
-//     x: endRightX,
-//     y: getEndTopY(),
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#diy", {
-//     x: endRightX,
-//     y: endBottomY,
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#motion", {
-//     x: endLeftX,
-//     y: endBottomY,
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-// }
-
-
-// export function updateDimensionsNoMargins() {
-//   setTimeout(() => {
-//     const thumbWidth = Math.min(300, window.innerWidth / 6);
-//     console.log("thumbWidth: ", thumbWidth);
-//     const screenWidthHalved = window.innerWidth / 2;
-//     const screenHeightHalved = window.innerHeight / 2;
-//     const widthThumb = getThumbWidthWithoutMargin();
-//     const marginWidth = getThumbMargin();
-
-//     updateEndTopY();
-
-//     const updatedEndTopY = getEndTopY(); // ✅ Fetch dynamically
-//     if (updatedEndTopY === undefined) {
-//       console.error("Error: getEndTopY() returned undefined!");
-//       return; 
-//     }
-
-//     const endTopY = updatedEndTopY + marginWidth - 15; 
-//     const endLeftX = screenWidthHalved - (widthThumb + marginWidth);
-//     const endRightX = screenWidthHalved - marginWidth;
-//     const endBottomY = updateEndBottomY() - marginWidth - 15;
-
-//     collectThumbs();
-//     updateModalDimensions(endTopY);
-//     formControl(endTopY);
-//   }, 450); // delay to ensure scrollbar removal takes effect
-// }
-
-// export function updateEndBottomY() {
-//   if (
-//     window.matchMedia("(orientation: landscape) and (max-width: 991.98px)")
-//       .matches
-//   ) {
-//     endBottomY = window.innerHeight * 1.275 + totalThumbWidth; // Adjust multiplier for this condition
-//   } else {
-//     endBottomY = window.innerHeight * 1.325 + totalThumbWidth;
-//   }
-//   return endBottomY;
-// }
-
-// export function collectThumbs() {
-//   gsap.to(domElements.thumbElements[0], {
-//     scale: 1,
-//     x: endLeftX,
-//     y: getEndTopY(),
-//     duration: 1,
-//     ease: "power2.out",
-//   }); console.log("collectThumbs-navbar 467 software : ", domElements.thumbElements[0] + " endLeftX: " + endLeftX + "getEndTopY() " + getEndTopY() + " MOTION: " + motion);
-//   gsap.to(photography, {
-//     scale: 1,
-//     x: endRightX,
-//     y: getEndTopY(),
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#diy", {
-//     scale: 1,
-//     x: endRightX,
-//     y: endBottomY,
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#motion", {
-//     scale: 1,
-//     x: endLeftX,
-//     y: endBottomY,
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-// }
-
-
-
-// export function updateEndTopY() {
-//   if (window.matchMedia("(orientation: landscape) and (max-width: 991.98px)").matches) {
-//     endTopY = window.innerHeight * 1.275;
-//   } else {
-//     endTopY = window.innerHeight * 1.325;
-//   }
-//   return endTopY;
-// }
-
-// export function getEndTopY() {
-//   return endTopY;
-// }
